feat: Enhance catalog and reviews pages with new features and styles
- Added product click handling in CatalogPage to navigate to product details.
- Implemented custom animations in globals.css for smoother transitions.
- Redesigned Home page with a hero section, services, and portfolio display.
- Improved ReviewsPage with modal image viewing and ordering functionality.
- Updated PhotoReview interface in database types for better data structure.
- Fixed price validation to ensure prices are parsed as integers.

+++ src/app/api/products/route.ts
@@ -53,7 +53,7 @@ export async function POST(request: NextRequest) {
     const productData = {
       name: body.name?.trim(),
       description: body.description?.trim(),
-      price: parseFloat(body.price),
+      price: parseInt(body.price),
       category: body.category,
       images: body.images || [],
       available: body.available !== false,
